pyproject.toml
[tool.poetry]
name = "SmartServe"
version = "0.0.1"
description = ""
authors = ["CarrotManMatt <matt@carrotmanmatt.com>"]
keywords = ["django-rest-framework"]
classifiers = [
    "Private :: Do Not Upload",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Typing :: Typed",
    "Topic :: Internet"
]

[tool.poetry.dependencies]
python = "^3.11.3"
django = "4.2.1"
djangorestframework = "^3.14.0"
django-rest-knox = "^4.2.0"
django-environ = "^0.10.0"
django-admin-rangefilter = "^0.10.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.3"
django-stubs = {extras = ["compatible-mypy"], version = "4.2.1"}
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^3.14.0"}

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
warn_unused_configs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_def = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true
strict_equality = true
strict_concatenate = true
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "core.settings"
